llm-monitoring-system/pyproject.toml
[project]
name = "llm-monitoring-system"
version = "0.1.0"
description = "LLM Monitoring System for Governmental Organizations"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "schedule>=1.2.0",
    "openai>=1.3.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.0",
    "httpx>=0.25.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0"
]
